+++ voucher-impl/src/main/java/com/dwarfeng/voucher/impl/cache/CheckerInfoCacheImpl.java
@@ -1,11 +1,10 @@
 package com.dwarfeng.voucher.impl.cache;
 
-import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.cache.BatchBaseCache;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
-import com.dwarfeng.voucher.sdk.bean.entity.FastJsonCheckerInfo;
 import com.dwarfeng.voucher.stack.bean.entity.CheckerInfo;
 import com.dwarfeng.voucher.stack.cache.CheckerInfoCache;
 import org.springframework.stereotype.Repository;
@@ -16,62 +15,61 @@ import java.util.List;
 @Repository
 public class CheckerInfoCacheImpl implements CheckerInfoCache {
 
-    private final RedisBatchBaseCache<StringIdKey, CheckerInfo, FastJsonCheckerInfo>
-            checkerInfoBatchBaseDelegate;
+    private final BatchBaseCache<StringIdKey, CheckerInfo> batchBaseCache;
 
     public CheckerInfoCacheImpl(
-            RedisBatchBaseCache<StringIdKey, CheckerInfo, FastJsonCheckerInfo> checkerInfoBatchBaseDelegate
+            BatchBaseCache<StringIdKey, CheckerInfo> batchBaseCache
     ) {
-        this.checkerInfoBatchBaseDelegate = checkerInfoBatchBaseDelegate;
+        this.batchBaseCache = batchBaseCache;
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean exists(StringIdKey key) throws CacheException {
-        return checkerInfoBatchBaseDelegate.exists(key);
+        return batchBaseCache.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public CheckerInfo get(StringIdKey key) throws CacheException {
-        return checkerInfoBatchBaseDelegate.get(key);
+        return batchBaseCache.get(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void push(CheckerInfo value, long timeout) throws CacheException {
-        checkerInfoBatchBaseDelegate.push(value, timeout);
+        batchBaseCache.push(value, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void delete(StringIdKey key) throws CacheException {
-        checkerInfoBatchBaseDelegate.delete(key);
+        batchBaseCache.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void clear() throws CacheException {
-        checkerInfoBatchBaseDelegate.clear();
+        batchBaseCache.clear();
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean allExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
-        return checkerInfoBatchBaseDelegate.allExists(keys);
+        return batchBaseCache.allExists(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
-        return checkerInfoBatchBaseDelegate.nonExists(keys);
+        return batchBaseCache.nonExists(keys);
     }
 
     @Override
@@ -79,20 +77,20 @@ public class CheckerInfoCacheImpl implements CheckerInfoCache {
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public List<CheckerInfo> batchGet(@SkipRecord List<StringIdKey> keys) throws CacheException {
-        return checkerInfoBatchBaseDelegate.batchGet(keys);
+        return batchBaseCache.batchGet(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void batchPush(@SkipRecord List<CheckerInfo> entities, long timeout) throws CacheException {
-        checkerInfoBatchBaseDelegate.batchPush(entities, timeout);
+        batchBaseCache.batchPush(entities, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void batchDelete(@SkipRecord List<StringIdKey> keys) throws CacheException {
-        checkerInfoBatchBaseDelegate.batchDelete(keys);
+        batchBaseCache.batchDelete(keys);
     }
 }