+++ voucher-impl/src/main/java/com/dwarfeng/voucher/impl/cache/VoucherCategoryCacheImpl.java
@@ -1,11 +1,10 @@
 package com.dwarfeng.voucher.impl.cache;
 
-import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.cache.BatchBaseCache;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
-import com.dwarfeng.voucher.sdk.bean.entity.FastJsonVoucherCategory;
 import com.dwarfeng.voucher.stack.bean.entity.VoucherCategory;
 import com.dwarfeng.voucher.stack.cache.VoucherCategoryCache;
 import org.springframework.stereotype.Repository;
@@ -16,62 +15,61 @@ import java.util.List;
 @Repository
 public class VoucherCategoryCacheImpl implements VoucherCategoryCache {
 
-    private final RedisBatchBaseCache<StringIdKey, VoucherCategory, FastJsonVoucherCategory>
-            voucherCategoryBatchBaseDelegate;
+    private final BatchBaseCache<StringIdKey, VoucherCategory> batchBaseCache;
 
     public VoucherCategoryCacheImpl(
-            RedisBatchBaseCache<StringIdKey, VoucherCategory, FastJsonVoucherCategory> voucherCategoryBatchBaseDelegate
+            BatchBaseCache<StringIdKey, VoucherCategory> batchBaseCache
     ) {
-        this.voucherCategoryBatchBaseDelegate = voucherCategoryBatchBaseDelegate;
+        this.batchBaseCache = batchBaseCache;
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean exists(StringIdKey key) throws CacheException {
-        return voucherCategoryBatchBaseDelegate.exists(key);
+        return batchBaseCache.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public VoucherCategory get(StringIdKey key) throws CacheException {
-        return voucherCategoryBatchBaseDelegate.get(key);
+        return batchBaseCache.get(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void push(VoucherCategory value, long timeout) throws CacheException {
-        voucherCategoryBatchBaseDelegate.push(value, timeout);
+        batchBaseCache.push(value, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void delete(StringIdKey key) throws CacheException {
-        voucherCategoryBatchBaseDelegate.delete(key);
+        batchBaseCache.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void clear() throws CacheException {
-        voucherCategoryBatchBaseDelegate.clear();
+        batchBaseCache.clear();
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean allExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
-        return voucherCategoryBatchBaseDelegate.allExists(keys);
+        return batchBaseCache.allExists(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
-        return voucherCategoryBatchBaseDelegate.nonExists(keys);
+        return batchBaseCache.nonExists(keys);
     }
 
     @Override
@@ -79,20 +77,20 @@ public class VoucherCategoryCacheImpl implements VoucherCategoryCache {
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public List<VoucherCategory> batchGet(@SkipRecord List<StringIdKey> keys) throws CacheException {
-        return voucherCategoryBatchBaseDelegate.batchGet(keys);
+        return batchBaseCache.batchGet(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void batchPush(@SkipRecord List<VoucherCategory> entities, long timeout) throws CacheException {
-        voucherCategoryBatchBaseDelegate.batchPush(entities, timeout);
+        batchBaseCache.batchPush(entities, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void batchDelete(@SkipRecord List<StringIdKey> keys) throws CacheException {
-        voucherCategoryBatchBaseDelegate.batchDelete(keys);
+        batchBaseCache.batchDelete(keys);
     }
 }