+++ voucher-impl/src/main/java/com/dwarfeng/voucher/impl/cache/VoucherVariableCacheImpl.java
@@ -1,10 +1,9 @@
 package com.dwarfeng.voucher.impl.cache;
 
-import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.stack.cache.BatchBaseCache;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
-import com.dwarfeng.voucher.sdk.bean.entity.FastJsonVoucherVariable;
 import com.dwarfeng.voucher.stack.bean.entity.VoucherVariable;
 import com.dwarfeng.voucher.stack.bean.key.VoucherVariableKey;
 import com.dwarfeng.voucher.stack.cache.VoucherVariableCache;
@@ -16,63 +15,61 @@ import java.util.List;
 @Repository
 public class VoucherVariableCacheImpl implements VoucherVariableCache {
 
-    private final RedisBatchBaseCache<VoucherVariableKey, VoucherVariable, FastJsonVoucherVariable>
-            voucherVariableBatchBaseDelegate;
+    private final BatchBaseCache<VoucherVariableKey, VoucherVariable> batchBaseCache;
 
     public VoucherVariableCacheImpl(
-            RedisBatchBaseCache<VoucherVariableKey, VoucherVariable, FastJsonVoucherVariable>
-                    voucherVariableBatchBaseDelegate
+            BatchBaseCache<VoucherVariableKey, VoucherVariable> batchBaseCache
     ) {
-        this.voucherVariableBatchBaseDelegate = voucherVariableBatchBaseDelegate;
+        this.batchBaseCache = batchBaseCache;
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean exists(VoucherVariableKey key) throws CacheException {
-        return voucherVariableBatchBaseDelegate.exists(key);
+        return batchBaseCache.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public VoucherVariable get(VoucherVariableKey key) throws CacheException {
-        return voucherVariableBatchBaseDelegate.get(key);
+        return batchBaseCache.get(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void push(VoucherVariable value, long timeout) throws CacheException {
-        voucherVariableBatchBaseDelegate.push(value, timeout);
+        batchBaseCache.push(value, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void delete(VoucherVariableKey key) throws CacheException {
-        voucherVariableBatchBaseDelegate.delete(key);
+        batchBaseCache.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void clear() throws CacheException {
-        voucherVariableBatchBaseDelegate.clear();
+        batchBaseCache.clear();
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean allExists(@SkipRecord List<VoucherVariableKey> keys) throws CacheException {
-        return voucherVariableBatchBaseDelegate.allExists(keys);
+        return batchBaseCache.allExists(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean nonExists(@SkipRecord List<VoucherVariableKey> keys) throws CacheException {
-        return voucherVariableBatchBaseDelegate.nonExists(keys);
+        return batchBaseCache.nonExists(keys);
     }
 
     @Override
@@ -80,20 +77,20 @@ public class VoucherVariableCacheImpl implements VoucherVariableCache {
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public List<VoucherVariable> batchGet(@SkipRecord List<VoucherVariableKey> keys) throws CacheException {
-        return voucherVariableBatchBaseDelegate.batchGet(keys);
+        return batchBaseCache.batchGet(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void batchPush(@SkipRecord List<VoucherVariable> entities, long timeout) throws CacheException {
-        voucherVariableBatchBaseDelegate.batchPush(entities, timeout);
+        batchBaseCache.batchPush(entities, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void batchDelete(@SkipRecord List<VoucherVariableKey> keys) throws CacheException {
-        voucherVariableBatchBaseDelegate.batchDelete(keys);
+        batchBaseCache.batchDelete(keys);
     }
 }