+++ voucher-impl/src/main/java/com/dwarfeng/voucher/impl/dao/CheckerInfoDaoImpl.java
@@ -1,15 +1,13 @@
 package com.dwarfeng.voucher.impl.dao;
 
-import com.dwarfeng.subgrade.impl.dao.HibernateBatchBaseDao;
-import com.dwarfeng.subgrade.impl.dao.HibernateEntireLookupDao;
-import com.dwarfeng.subgrade.impl.dao.HibernatePresetLookupDao;
-import com.dwarfeng.subgrade.sdk.bean.key.HibernateStringIdKey;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.dao.BatchBaseDao;
+import com.dwarfeng.subgrade.stack.dao.EntireLookupDao;
+import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
-import com.dwarfeng.voucher.impl.bean.entity.HibernateCheckerInfo;
 import com.dwarfeng.voucher.stack.bean.entity.CheckerInfo;
 import com.dwarfeng.voucher.stack.dao.CheckerInfoDao;
 import org.springframework.stereotype.Repository;
@@ -20,16 +18,14 @@ import java.util.List;
 @Repository
 public class CheckerInfoDaoImpl implements CheckerInfoDao {
 
-    private final HibernateBatchBaseDao<StringIdKey, HibernateStringIdKey, CheckerInfo, HibernateCheckerInfo>
-            batchBaseDao;
-    private final HibernateEntireLookupDao<CheckerInfo, HibernateCheckerInfo> entireLookupDao;
-    private final HibernatePresetLookupDao<CheckerInfo, HibernateCheckerInfo> presetLookupDao;
+    private final BatchBaseDao<StringIdKey, CheckerInfo> batchBaseDao;
+    private final EntireLookupDao<CheckerInfo> entireLookupDao;
+    private final PresetLookupDao<CheckerInfo> presetLookupDao;
 
     public CheckerInfoDaoImpl(
-            HibernateBatchBaseDao<StringIdKey, HibernateStringIdKey, CheckerInfo, HibernateCheckerInfo>
-                    batchBaseDao,
-            HibernateEntireLookupDao<CheckerInfo, HibernateCheckerInfo> entireLookupDao,
-            HibernatePresetLookupDao<CheckerInfo, HibernateCheckerInfo> presetLookupDao
+            BatchBaseDao<StringIdKey, CheckerInfo> batchBaseDao,
+            EntireLookupDao<CheckerInfo> entireLookupDao,
+            PresetLookupDao<CheckerInfo> presetLookupDao
     ) {
         this.batchBaseDao = batchBaseDao;
         this.entireLookupDao = entireLookupDao;