+++ voucher-impl/src/main/java/com/dwarfeng/voucher/impl/dao/CheckerSupportDaoImpl.java
@@ -1,15 +1,13 @@
 package com.dwarfeng.voucher.impl.dao;
 
-import com.dwarfeng.subgrade.impl.dao.HibernateBatchBaseDao;
-import com.dwarfeng.subgrade.impl.dao.HibernateEntireLookupDao;
-import com.dwarfeng.subgrade.impl.dao.HibernatePresetLookupDao;
-import com.dwarfeng.subgrade.sdk.bean.key.HibernateStringIdKey;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.dao.BatchBaseDao;
+import com.dwarfeng.subgrade.stack.dao.EntireLookupDao;
+import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
-import com.dwarfeng.voucher.impl.bean.entity.HibernateCheckerSupport;
 import com.dwarfeng.voucher.stack.bean.entity.CheckerSupport;
 import com.dwarfeng.voucher.stack.dao.CheckerSupportDao;
 import org.springframework.stereotype.Repository;
@@ -20,16 +18,14 @@ import java.util.List;
 @Repository
 public class CheckerSupportDaoImpl implements CheckerSupportDao {
 
-    private final HibernateBatchBaseDao<StringIdKey, HibernateStringIdKey, CheckerSupport, HibernateCheckerSupport>
-            batchBaseDao;
-    private final HibernateEntireLookupDao<CheckerSupport, HibernateCheckerSupport> entireLookupDao;
-    private final HibernatePresetLookupDao<CheckerSupport, HibernateCheckerSupport> presetLookupDao;
+    private final BatchBaseDao<StringIdKey, CheckerSupport> batchBaseDao;
+    private final EntireLookupDao<CheckerSupport> entireLookupDao;
+    private final PresetLookupDao<CheckerSupport> presetLookupDao;
 
     public CheckerSupportDaoImpl(
-            HibernateBatchBaseDao<StringIdKey, HibernateStringIdKey, CheckerSupport, HibernateCheckerSupport>
-                    batchBaseDao,
-            HibernateEntireLookupDao<CheckerSupport, HibernateCheckerSupport> entireLookupDao,
-            HibernatePresetLookupDao<CheckerSupport, HibernateCheckerSupport> presetLookupDao
+            BatchBaseDao<StringIdKey, CheckerSupport> batchBaseDao,
+            EntireLookupDao<CheckerSupport> entireLookupDao,
+            PresetLookupDao<CheckerSupport> presetLookupDao
     ) {
         this.batchBaseDao = batchBaseDao;
         this.entireLookupDao = entireLookupDao;