+++ voucher-impl/src/main/java/com/dwarfeng/voucher/impl/dao/VoucherVariableDaoImpl.java
@@ -1,14 +1,12 @@
 package com.dwarfeng.voucher.impl.dao;
 
-import com.dwarfeng.subgrade.impl.dao.HibernateBatchBaseDao;
-import com.dwarfeng.subgrade.impl.dao.HibernateEntireLookupDao;
-import com.dwarfeng.subgrade.impl.dao.HibernatePresetLookupDao;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.dao.BatchBaseDao;
+import com.dwarfeng.subgrade.stack.dao.EntireLookupDao;
+import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
-import com.dwarfeng.voucher.impl.bean.entity.HibernateVoucherVariable;
-import com.dwarfeng.voucher.impl.bean.key.HibernateVoucherVariableKey;
 import com.dwarfeng.voucher.stack.bean.entity.VoucherVariable;
 import com.dwarfeng.voucher.stack.bean.key.VoucherVariableKey;
 import com.dwarfeng.voucher.stack.dao.VoucherVariableDao;
@@ -20,16 +18,14 @@ import java.util.List;
 @Repository
 public class VoucherVariableDaoImpl implements VoucherVariableDao {
 
-    private final HibernateBatchBaseDao<VoucherVariableKey, HibernateVoucherVariableKey, VoucherVariable,
-            HibernateVoucherVariable> batchBaseDao;
-    private final HibernateEntireLookupDao<VoucherVariable, HibernateVoucherVariable> entireLookupDao;
-    private final HibernatePresetLookupDao<VoucherVariable, HibernateVoucherVariable> presetLookupDao;
+    private final BatchBaseDao<VoucherVariableKey, VoucherVariable> batchBaseDao;
+    private final EntireLookupDao<VoucherVariable> entireLookupDao;
+    private final PresetLookupDao<VoucherVariable> presetLookupDao;
 
     public VoucherVariableDaoImpl(
-            HibernateBatchBaseDao<VoucherVariableKey, HibernateVoucherVariableKey, VoucherVariable,
-                    HibernateVoucherVariable> batchBaseDao,
-            HibernateEntireLookupDao<VoucherVariable, HibernateVoucherVariable> entireLookupDao,
-            HibernatePresetLookupDao<VoucherVariable, HibernateVoucherVariable> presetLookupDao
+            BatchBaseDao<VoucherVariableKey, VoucherVariable> batchBaseDao,
+            EntireLookupDao<VoucherVariable> entireLookupDao,
+            PresetLookupDao<VoucherVariable> presetLookupDao
     ) {
         this.batchBaseDao = batchBaseDao;
         this.entireLookupDao = entireLookupDao;