+++ voucher-impl/src/main/java/com/dwarfeng/voucher/impl/service/VoucherVariableMaintainServiceImpl.java
@@ -1,13 +1,13 @@
 package com.dwarfeng.voucher.impl.service;
 
-import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
-import com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService;
-import com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import com.dwarfeng.subgrade.stack.service.BatchCrudService;
+import com.dwarfeng.subgrade.stack.service.EntireLookupService;
+import com.dwarfeng.subgrade.stack.service.PresetLookupService;
 import com.dwarfeng.voucher.stack.bean.entity.VoucherVariable;
 import com.dwarfeng.voucher.stack.bean.key.VoucherVariableKey;
 import com.dwarfeng.voucher.stack.service.VoucherVariableMaintainService;
@@ -19,14 +19,14 @@ import java.util.List;
 @Component
 public class VoucherVariableMaintainServiceImpl implements VoucherVariableMaintainService {
 
-    private final GeneralBatchCrudService<VoucherVariableKey, VoucherVariable> crudService;
-    private final DaoOnlyEntireLookupService<VoucherVariable> entireLookupService;
-    private final DaoOnlyPresetLookupService<VoucherVariable> presetLookupService;
+    private final BatchCrudService<VoucherVariableKey, VoucherVariable> crudService;
+    private final EntireLookupService<VoucherVariable> entireLookupService;
+    private final PresetLookupService<VoucherVariable> presetLookupService;
 
     public VoucherVariableMaintainServiceImpl(
-            GeneralBatchCrudService<VoucherVariableKey, VoucherVariable> crudService,
-            DaoOnlyEntireLookupService<VoucherVariable> entireLookupService,
-            DaoOnlyPresetLookupService<VoucherVariable> presetLookupService
+            BatchCrudService<VoucherVariableKey, VoucherVariable> crudService,
+            EntireLookupService<VoucherVariable> entireLookupService,
+            PresetLookupService<VoucherVariable> presetLookupService
     ) {
         this.crudService = crudService;
         this.entireLookupService = entireLookupService;